requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
alembic==1.7.5
    # via -r requirements.in
anyio==3.5.0
    # via starlette
asgiref==3.5.0
    # via uvicorn
backports.cached-property==1.0.1
    # via strawberry-graphql
click==8.0.3
    # via
    #   strawberry-graphql
    #   uvicorn
fastapi==0.73.0
    # via
    #   -r requirements.in
    #   strawberry-graphql
graphql-core==3.1.7
    # via strawberry-graphql
h11==0.13.0
    # via uvicorn
httptools==0.3.0
    # via uvicorn
idna==3.3
    # via anyio
importlib-metadata==4.10.1
    # via alembic
importlib-resources==5.4.0
    # via alembic
mako==1.1.6
    # via alembic
markupsafe==2.0.1
    # via mako
psycopg2==2.9.3
    # via -r requirements.in
pydantic==1.9.0
    # via fastapi
pygments==2.11.2
    # via strawberry-graphql
python-dateutil==2.8.2
    # via strawberry-graphql
python-dotenv==0.19.2
    # via uvicorn
python-multipart==0.0.5
    # via strawberry-graphql
pyyaml==6.0
    # via uvicorn
sentinel==0.3.0
    # via strawberry-graphql
six==1.16.0
    # via
    #   python-dateutil
    #   python-multipart
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.31
    # via
    #   -r requirements.in
    #   alembic
starlette==0.17.1
    # via fastapi
strawberry-graphql[fastapi]==0.95.1
    # via -r requirements.in
typing-extensions==4.0.1
    # via
    #   pydantic
    #   strawberry-graphql
uvicorn[standard]==0.17.0.post1
    # via -r requirements.in
uvloop==0.16.0
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==10.1
    # via uvicorn
zipp==3.7.0
    # via
    #   importlib-metadata
    #   importlib-resources
